\name{pairheatmap-package}
\alias{pairheatmap-package}
\docType{package}
\title{
A tool for comparing heatmaps
}
\description{
A tool to compare two heatmaps and discover patterns within and across groups. In the context of biology, group can be defined based on gene ontology.
}
\details{
\tabular{ll}{
Package: \tab pairheatmap\cr
Type: \tab Package\cr
Version: \tab 1.0.0 \cr
Date: \tab 2011-12-27 \cr
License: \tab What license is it under?\cr
}
The main interface is pairheatmap() function
}
\author{
Xiaoyong Sun
Maintainer: johnsunx1<johnsunx1@gmail.com>
}

\keyword{ package }

\examples{
pairheatmap(mtcars, mtcars)
}
